+++ src/util/database/SQLiteDB.java
@@ -268,7 +268,7 @@ public class SQLiteDB extends SQLDB{
         db.readFlushSetup(XMLtools.getFirstChildByTag(dbe, "flush"));
 
         // How many seconds before the connection is considered idle (and closed)
-        db.idleTime = TimeTools.parsePeriodStringToSeconds(XMLtools.getChildValueByTag(dbe,"idleclose","5m"));
+        db.idleTime = (int)TimeTools.parsePeriodStringToSeconds(XMLtools.getChildValueByTag(dbe,"idleclose","5m"));
 
         /* Views */
         for( Element view : XMLtools.getChildElements(dbe, "view")){
